window.open with noopener shouldn't create page in opener process https://bugs.webkit.org/show_bug.cgi?id=279031
Reviewed by NOBODY (OOPS!).

We used to make a page in the opener process then navigate it and immediately discard it.
That doesn't play well with site isolation's process accounting.  JS can't observe this
because window.open returns null in such a case.

In order to implement this, I need to only call API::PageConfiguration::setRelatedPage
if there is actually an opener, and I need to remove the now duplicate checks that verify
the API client is using the provided WKWebViewConfiguration by checking the related page.
There is already another check that verifies using the opener info.

To get iframe sandboxing to keep working, I had to send the NavigationActionData to
WebPageProxy::loadRequest and serialize the sandbox flags from WebChromeClient::createWindow
to the UI process.  Navigation::lastNavigationAction is where those flags are sent to
the new process from, and this introduces a new use of WebPageProxy::loadRequest to
start the load in a new process.

A few layout tests in http/tests/dom count the number of windows in the process that is
finishing the test.  Since one of the windows is now opening in another process in modern
WebKit, the count goes from 3 to 2.  WebKitLegacy test expectations are unchanged for
those three tests.

* Source/WebCore/page/LocalDOMWindow.cpp:
(WebCore::LocalDOMWindow::createWindow):
* Source/WebKit/UIProcess/Cocoa/UIDelegate.mm:
(WebKit::UIDelegate::UIClient::createNewPage):
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::createNewPage):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/ProcessSwapOnNavigation.mm:
((ProcessSwap, SameSiteWindowOpenNoOpener)):
((ProcessSwap, SameSiteWindowOpenWithNameNoOpener)):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/SiteIsolation.mm:
(TestWebKitAPI::TEST(SiteIsolation, OpenWithNoopener)):