Fix http/tests/security/aboutBlank/security-context-window-open.html with site isolation enabled https://bugs.webkit.org/show_bug.cgi?id=284343 rdar://141188809
Reviewed by NOBODY (OOPS!).

In Document::initSecurityContext we are setting the cookie URL based on the owner frame.
Before this PR we were assuming the owner frame was always a LocalFrame, and it is with
site isolation off.  With site isolation on, though, we need to not return early if there's
a RemoteFrame and use the Page's main frame URL instead, which is updated in all processes.
There's still more work to do in this area, but this makes the test pass instead of crashing.

* LayoutTests/platform/mac-site-isolation/TestExpectations:
* Source/WebCore/dom/Document.cpp:
(WebCore::Document::initSecurityContext):