Add a RefPtr constructor that takes a reference https://bugs.webkit.org/show_bug.cgi?id=288668 rdar://145703009
Reviewed by NOBODY (OOPS!).

You can make a WeakPtr from a T&, why not a RefPtr?
Usually you want a Ref when you're doing this, but sometimes,
like in DocumentFullscreen, the type has to be nullable but start
out non-null.  This avoids an unnecessary null check, and it
makes the code more elegant.

* Source/WTF/wtf/RefPtr.h:
(WTF::RefPtr::RefPtr):
* Source/WebCore/dom/DocumentFullscreen.cpp:
(WebCore::documentsToUnfullscreen):