Use AuxiliaryProcessProxy::fromConnection instead of ProvisionalPageProxy::protectedProcess to get sending process https://bugs.webkit.org/show_bug.cgi?id=290503
Reviewed by NOBODY (OOPS!).

ProvisionalPageProxy::protectedProcess was used to answer the question "which process sent this message?"
for PSON.  For site isolation, we need something a little more robust.  We already have and use
AuxiliaryProcessProxy::fromConnection for this purpose, so startURLSchemeTaskShared is useless.

WebPageProxy::webPageIDInProcess already works correctly with site isolation enabled.  This PR makes
it work correctly when site isolation is not enabled and the PageIdentifier may be from the current
process or the next process, owned by the ProvisionalPageProxy.

* Source/WebKit/UIProcess/ProvisionalPageProxy.cpp:
(WebKit::ProvisionalPageProxy::didReceiveMessage):
(WebKit::ProvisionalPageProxy::startURLSchemeTask): Deleted.
* Source/WebKit/UIProcess/ProvisionalPageProxy.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::startURLSchemeTask):
(WebKit::WebPageProxy::startURLSchemeTaskShared): Deleted.
* Source/WebKit/UIProcess/WebPageProxy.h: