Add a way to prepend event listeners in autofill content worlds https://bugs.webkit.org/show_bug.cgi?id=304652 rdar://167101754
Reviewed by NOBODY (OOPS!).

Safari's autofill implementation runs some JS that adds event listeners before anything else can.
This adds a way to add an event listener and bump it to the front of the line to be called before
other event listeners.  This should remove the need to run JS as early as we do.
This is limited to autofill content worlds because everyone wants to cut in line, but the normal
worlds' JS needs to play fair like it has.
I initially put the function prependEventListener on DOMWindow because that's the only place we
currently use it and it seemed less expensive than putting it on EventTarget, which is instantiated
many times per global object.  We may be able to revisit that if needed.

Test: Tools/TestWebKitAPI/Tests/mac/KeyboardEventTests.mm

* Source/WebCore/bindings/js/WebCoreBuiltinNames.h:
* Source/WebCore/bindings/scripts/CodeGeneratorJS.pm:
(GenerateHeader):
* Source/WebCore/bindings/scripts/test/JS/JSDOMWindow.h:
* Source/WebCore/dom/EventListenerMap.cpp:
(WebCore::EventListenerMap::moveLastListenerToFirst):
* Source/WebCore/dom/EventListenerMap.h:
* Source/WebCore/dom/EventTarget.cpp:
(WebCore::EventTarget::prependEventListener):
* Source/WebCore/dom/EventTarget.h:
* Source/WebCore/page/DOMWindow.idl:
* Tools/TestWebKitAPI/Tests/mac/KeyboardEventTests.mm:
(TestWebKitAPI::TEST(KeyboardEventTests, PrependEventListener)):